Refactoring and test coverage fix
